Updated Tiffin Config Feature
We chose to use the Bootstrap Grid for the layout that Masonry was originally used. With Bootstrap Grid now used exclusively, there is no need to keep Masonry and the overhead incurred in loading the module and library files on every page load. Run updatedb in drush to have it disable and uninstall the masonry, masonry_panels, and masonry_views modules.